+++ src/main/java/com/baosong/supplyme/service/impl/SupplierServiceImpl.java
@@ -1,10 +1,20 @@
 package com.baosong.supplyme.service.impl;
 
-import com.baosong.supplyme.service.SupplierService;
-import com.baosong.supplyme.service.UserService;
+import static org.elasticsearch.index.query.QueryBuilders.queryStringQuery;
+
+import java.time.Instant;
+import java.util.Optional;
+
 import com.baosong.supplyme.domain.Supplier;
+import com.baosong.supplyme.domain.errors.ServiceException;
 import com.baosong.supplyme.repository.SupplierRepository;
 import com.baosong.supplyme.repository.search.SupplierSearchRepository;
+import com.baosong.supplyme.security.AuthoritiesConstants;
+import com.baosong.supplyme.security.SecurityUtils;
+import com.baosong.supplyme.service.MutablePropertiesService;
+import com.baosong.supplyme.service.SupplierService;
+import com.baosong.supplyme.service.UserService;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,11 +22,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.time.Instant;
-import java.util.Optional;
-
-import static org.elasticsearch.index.query.QueryBuilders.*;
 
 /**
  * Service Implementation for managing Supplier.
@@ -34,6 +39,9 @@ public class SupplierServiceImpl implements SupplierService {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private MutablePropertiesService mutablePropertiesService;
+
     public SupplierServiceImpl(SupplierRepository supplierRepository, SupplierSearchRepository supplierSearchRepository) {
         this.supplierRepository = supplierRepository;
         this.supplierSearchRepository = supplierSearchRepository;
@@ -46,11 +54,27 @@ public class SupplierServiceImpl implements SupplierService {
      * @return the persisted entity
      */
     @Override
-    public Supplier save(Supplier supplier) {
+    public Supplier save(Supplier supplier) throws ServiceException {
         log.debug("Request to save Supplier : {}", supplier);
+        if (!SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.SUPPLIER_MANAGER)) {
+            throw new ServiceException(
+                String.format("The current user cannot save Supplier %s", supplier.getReferenceNumber()),
+                "supplier.edit.forbidden");
+        }
         if (supplier.getId() == null) {
             supplier.creationDate(Instant.now()).setCreationUser(userService.getCurrentUser().get());
+            if (supplier.isTemporary()) {
+                supplier.setReferenceNumber(this.mutablePropertiesService.getNewTemporarySupplierCode());
+            }
         }
+        Supplier sameReferenceNumberSupplier = this.supplierRepository.getByReferenceNumber(supplier.getReferenceNumber());
+        if (sameReferenceNumberSupplier != null && !sameReferenceNumberSupplier.getId().equals(supplier.getId())) {
+            throw new ServiceException(
+                String.format("Ref %s is already used by Supplier %d.", supplier.getReferenceNumber(), supplier.getId()),
+                "supplier.edit.alreadyExists"
+            );
+        }
+
         Supplier result = supplierRepository.save(supplier);
         supplierSearchRepository.save(result);
         return result;
@@ -87,10 +111,16 @@ public class SupplierServiceImpl implements SupplierService {
      * Delete the supplier by id.
      *
      * @param id the id of the entity
+     * @throws ServiceException
      */
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws ServiceException {
         log.debug("Request to delete Supplier : {}", id);
+        if (!SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.SUPPLIER_MANAGER)) {
+            throw new ServiceException(
+                String.format("The current user cannot delete Supplier %d", id),
+                "supplier.delete.forbidden");
+        }
         supplierRepository.deleteById(id);
         supplierSearchRepository.deleteById(id);
     }
